Xtra

+++ app/controllers/registered_applications_controller.rb
@@ -1,19 +1,22 @@
 #
 class RegisteredApplicationsController < ApplicationController
   before_action :make_view_helper
+  before_filter :authenticate_user!
 
   # A frequent practice is to place the standard CRUD actions in each controller
   # in the following order:
   #   index, show, new, edit, create, update and destroy.
   #
   def index
-    @view.apps = RegisteredApplication.all
+    @view.apps = RegisteredApplication.where('user_id = ?', current_user.id)
+    authorize @view.apps
     # Response: Controller will forward_to
     #           /views/registered_applications/index.html.erb with @view
   end
 
   def show
     @view.app = RegisteredApplication.find(params[:id])
+    authorize @view.app
     # { ":bath:"=>1, ":football:"=>1, ":mahjong:"=>1, ":musical_keyboard:"=>1,
     #   ":saxophone:"=>1, ":slot_machine:"=>1
     # }
@@ -25,6 +28,7 @@ class RegisteredApplicationsController < ApplicationController
   end
 
   def new
+    authorize RegisteredApplication
     @view.app = RegisteredApplication.new
     # Response: Controller will forward_to
     #           /views/registered_applications/new.html.erb with @view
@@ -33,6 +37,7 @@ class RegisteredApplicationsController < ApplicationController
 
   # Note: we get here from a submit button via the new view.
   def create
+    authorize RegisteredApplication
     @view.app = RegisteredApplication.new(submitted_params_whitelist)
     @view.app.url = fixup_url
     @view.app.user = @view.current_user
@@ -49,6 +54,7 @@ class RegisteredApplicationsController < ApplicationController
 
   def edit
     @view.app = RegisteredApplication.find(params[:id])
+    authorize @view.app
     # Response: Controller will forward_to
     #           /views/registered_applications/edit.html.erb
     #           with @view.
@@ -58,6 +64,7 @@ class RegisteredApplicationsController < ApplicationController
   # Note: we get here from a submit button via the edit view.
   def update
     @view.app = RegisteredApplication.find(params[:id])
+    authorize @view.app
     @view.app.assign_attributes(submitted_params_whitelist)
 
     # Response: redirect to or forward_to a view with msgs
@@ -73,6 +80,7 @@ class RegisteredApplicationsController < ApplicationController
 
   def destroy
     @view.app = RegisteredApplication.find(params[:id])
+    authorize @view.app
     # Response: redirect to or forward_to to a view.
     if @view.app.destroy
       flash[:notice] = "\"#{@view.app.name}\" was deleted successfully."
@@ -82,6 +90,11 @@ class RegisteredApplicationsController < ApplicationController
       flash.now[:alert] = 'There was an error deleting this Registered Site.'
       render :show
     end
+  end
+
+  # Decide where back button takes us.
+  def page_back_button_path
+    registered_applications_path
   end
 
   private
